backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
itsdangerous==2.2.0
gunicorn==22.0.0
Werkzeug==3.0.3
Authlib==1.3.1
